handle complex call expressions (#440)
