+++ test/form/side-effect-n/_expected/amd.js
@@ -1,0 +1,13 @@
+define(function () { 'use strict';
+
+	function foo () {
+		console.log( 'foo' );
+	}
+
+	function bar () {
+		console.log( 'bar' );
+	}
+
+	( Math.random() < 0.5 ? foo : bar )();
+
+});
