+++ test/form/side-effect-n/_expected/cjs.js
@@ -1,0 +1,11 @@
+'use strict';
+
+function foo () {
+	console.log( 'foo' );
+}
+
+function bar () {
+	console.log( 'bar' );
+}
+
+( Math.random() < 0.5 ? foo : bar )();
