+++ test/form/side-effect-n/_expected/es6.js
@@ -1,0 +1,9 @@
+function foo () {
+	console.log( 'foo' );
+}
+
+function bar () {
+	console.log( 'bar' );
+}
+
+( Math.random() < 0.5 ? foo : bar )();
