+++ test/form/side-effect-n/_expected/iife.js
@@ -1,0 +1,14 @@
+(function () {
+	'use strict';
+
+	function foo () {
+		console.log( 'foo' );
+	}
+
+	function bar () {
+		console.log( 'bar' );
+	}
+
+	( Math.random() < 0.5 ? foo : bar )();
+
+}());
