+++ test/form/side-effect-o/_expected/cjs.js
@@ -1,0 +1,15 @@
+'use strict';
+
+function fn () {
+	return Math.random() < 0.5 ? foo : bar;
+}
+
+function foo () {
+	console.log( 'foo' );
+}
+
+function bar () {
+	console.log( 'bar' );
+}
+
+fn()();
