+++ test/form/side-effect-o/_expected/es6.js
@@ -1,0 +1,13 @@
+function fn () {
+	return Math.random() < 0.5 ? foo : bar;
+}
+
+function foo () {
+	console.log( 'foo' );
+}
+
+function bar () {
+	console.log( 'bar' );
+}
+
+fn()();
