+++ test/form/side-effect-o/_expected/iife.js
@@ -1,0 +1,18 @@
+(function () {
+	'use strict';
+
+	function fn () {
+		return Math.random() < 0.5 ? foo : bar;
+	}
+
+	function foo () {
+		console.log( 'foo' );
+	}
+
+	function bar () {
+		console.log( 'bar' );
+	}
+
+	fn()();
+
+}());
