+++ test/form/side-effect-o/_expected/umd.js
@@ -1,0 +1,21 @@
+(function (global, factory) {
+	typeof exports === 'object' && typeof module !== 'undefined' ? factory() :
+	typeof define === 'function' && define.amd ? define(factory) :
+	(factory());
+}(this, function () { 'use strict';
+
+	function fn () {
+		return Math.random() < 0.5 ? foo : bar;
+	}
+
+	function foo () {
+		console.log( 'foo' );
+	}
+
+	function bar () {
+		console.log( 'bar' );
+	}
+
+	fn()();
+
+}));
